chore: remove more unused client files and move used to new dir
