--- client/src/components/Input/EndpointMenu/PresetItems.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import PresetItem from './PresetItem';
-import type { TPreset } from 'librechat-data-provider';
-
-export default function PresetItems({ presets, onSelect, onChangePreset, onDeletePreset }) {
-  return (
-    <>
-      {presets.map((preset: TPreset) => (
-        <PresetItem
-          key={preset?.presetId ?? Math.random()}
-          value={preset}
-          onSelect={onSelect}
-          onChangePreset={onChangePreset}
-          onDeletePreset={onDeletePreset}
-          preset={preset}
-        />
-      ))}
-    </>
-  );
-}
