--- client/src/components/Input/EndpointMenu/index.ts
@@ -1,1 +1,0 @@
-export { default as EndpointMenu } from './EndpointMenu';
