+++ client/src/components/Input/SetKeyDialog/GoogleConfig.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { object, string } from 'zod';
 import { AuthKeys } from 'librechat-data-provider';
 import type { TConfigProps } from '~/common';
-import FileUpload from '~/components/Input/EndpointMenu/FileUpload';
+import FileUpload from '~/components/Chat/Input/Files/FileUpload';
 import { useLocalize, useMultipleKeys } from '~/hooks';
 import InputWithLabel from './InputWithLabel';
 import { Label } from '~/components/ui';
